Seperated core module into directory
